requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
deep-translator==1.11.4
filelock==3.16.1
fsspec==2024.10.0
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
googletrans==3.1.0a0
greenlet==3.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyee==12.0.0
pyparsing==3.2.0
python-dotenv==1.0.1
pytubefix==8.2.0
regex==2024.9.11
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
setuptools==75.3.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.66.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
youtube-transcript-api==0.6.2
